+++ lib/modulorails/railtie.rb
@@ -46,9 +46,6 @@ module Modulorails
       # Generate a template for CI/CD
       Modulorails.generate_ci_template
 
-      # Check database configuration
-      Modulorails.check_database_config
-
       # Add/update Rubocop config
       Modulorails.generate_rubocop_template